refactor: more sensible package names
